+++ src/main/kotlin/io/github/ayfri/kore/koreassistant/Constants.kt
@@ -1,19 +1,14 @@
 package io.github.ayfri.kore.koreassistant
 
 import com.intellij.openapi.util.IconLoader
-import io.github.ayfri.kore.koreassistant.gutter.DataPackGutterProvider
 import org.jetbrains.kotlin.name.FqName
-import org.jetbrains.kotlin.name.Name
 
-data object KoreNames {
-	val KORE_DATAPACK_CLASS_ID = FqName("io.github.ayfri.kore.dataPack")
-	val KORE_DATAPACK_NAME = Name.identifier("dataPack")
-	val KORE_FUNCTION_CLASS_ID = FqName("io.github.ayfri.kore.functions.function")
-	val KORE_FUNCTION_NAME = Name.identifier("function")
+object KoreNames {
+	val DATA_PACK = FqName("io.github.ayfri.kore.dataPack")
+	val FUNCTION = FqName("io.github.ayfri.kore.functions.function")
 }
 
-data object KoreIcons {
-	// Use java.classLoader to ensure compatibility across different environments
-	val KORE = IconLoader.getIcon("/images/kore-white.svg", DataPackGutterProvider::class.java.classLoader)
-	val FUNCTION = IconLoader.getIcon("/images/function.svg", DataPackGutterProvider::class.java.classLoader)
+object KoreIcons {
+	val KORE = IconLoader.getIcon("/images/kore-white.svg", KoreIcons::class.java.classLoader)
+	val FUNCTION = IconLoader.getIcon("/images/function.svg", KoreIcons::class.java.classLoader)
 }