+++ src/main/kotlin/io/github/ayfri/kore/koreassistant/gutter/DataPackGutterProvider.kt
@@ -4,8 +4,7 @@ import io.github.ayfri.kore.koreassistant.KoreIcons
 import io.github.ayfri.kore.koreassistant.KoreNames
 
 class DataPackGutterProvider : KoreCallGutterProvider(
-	shortName = KoreNames.KORE_DATAPACK_NAME,
-	fqName = KoreNames.KORE_DATAPACK_CLASS_ID,
+	fqName = KoreNames.DATA_PACK,
 	icon = KoreIcons.KORE,
 	tooltipTitle = "DataPack Definition",
 	tooltipText = "Kore dataPack definition",