+++ src/main/kotlin/io/github/ayfri/kore/koreassistant/gutter/FunctionGutterProvider.kt
@@ -4,8 +4,7 @@ import io.github.ayfri.kore.koreassistant.KoreIcons
 import io.github.ayfri.kore.koreassistant.KoreNames
 
 class FunctionGutterProvider : KoreCallGutterProvider(
-	shortName = KoreNames.KORE_FUNCTION_NAME,
-	fqName = KoreNames.KORE_FUNCTION_CLASS_ID,
+	fqName = KoreNames.FUNCTION,
 	icon = KoreIcons.FUNCTION,
 	tooltipTitle = "Function Definition",
 	tooltipText = "Kore function definition",